initiated lrdftransformers docs
