+++ suricate/lrdftransformers/__init__.py
@@ -1,5 +1,6 @@
-from suricate.lrdftransformers.base import LrDfTransformerMixin, cartesian_join
-from suricate.lrdftransformers.cartesian import CartesianLr, CartesianDataPasser, LrDfVisualHelper, create_lrdf_sbs
+from suricate.lrdftransformers.base import LrDfTransformerMixin
+from suricate.lrdftransformers.cartesian import CartesianLr, CartesianDataPasser, LrDfVisualHelper, create_lrdf_sbs, \
+    cartesian_join
 from suricate.lrdftransformers.exact import ExactConnector
 from suricate.lrdftransformers.fuzzy import FuzzyConnector
 from suricate.lrdftransformers.indexer import Indexer
